chore: Add agent starting status
Updated serverside validation for the agent status now means the
unspecified status is treated as invalid, so the cofide-agent fails to
start successfully as this is trying to set the status as unspecified
currently. This adds a new status for the agent to send when it is
starting instead of unspecified.