fix: Make role binding resource type a string
Role bindings cannot be attached to ephemeral types, it was a mistake to
originally include them in the enum here.

Have decided internally that using an enum here is potentially overly
restrictive going forwards, so using a string and documenting the
accepted values is more flexible (if sacrificing a little built-in type
safety).