chore: downgrade some dependencies
We were previously keenly updating dependencies using dependabot. This
forced consumers of this module to also update. In some circumstances
that is not preferable. As a library we should not force updates on our
consumers without good reason, so this change downgrades various
dependencies.

Also restrict dependabot updates to patch-level only.

Finally, enable dependabot updates for GitHub Actions.